# Conflicts:
#	engine/src/main/java/org/camunda/bpm/engine/impl/dmn/entity/repository/DecisionDefinitionQueryImpl.java

+++ engine/src/main/java/org/camunda/bpm/engine/impl/persistence/entity/HistoricProcessInstanceManager.java
@@ -19,10 +19,8 @@ package org.camunda.bpm.engine.impl.persistence.entity;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import org.camunda.bpm.engine.authorization.Resources;
 import org.camunda.bpm.engine.history.CleanableHistoricProcessInstanceReportResult;
@@ -38,6 +36,7 @@ import org.camunda.bpm.engine.impl.history.event.HistoricProcessInstanceEventEnt
 import org.camunda.bpm.engine.impl.interceptor.CommandContext;
 import org.camunda.bpm.engine.impl.persistence.AbstractHistoricManager;
 import org.camunda.bpm.engine.impl.util.ClockUtil;
+import org.camunda.bpm.engine.impl.util.ImmutablePair;
 
 /**
  * @author Tom Baeyens
@@ -71,19 +70,22 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
   }
 
   public void deleteHistoricProcessInstanceByIds(List<String> processInstanceIds) {
-    CommandContext commandContext = Context.getCommandContext();
+    if (isHistoryEnabled()) {
+      CommandContext commandContext = Context.getCommandContext();
 
-    commandContext.getHistoricDetailManager().deleteHistoricDetailsByProcessInstanceIds(processInstanceIds);
-    commandContext.getHistoricVariableInstanceManager().deleteHistoricVariableInstanceByProcessInstanceIds(processInstanceIds);
-    commandContext.getCommentManager().deleteCommentsByProcessInstanceIds(processInstanceIds);
-    commandContext.getAttachmentManager().deleteAttachmentsByProcessInstanceIds(processInstanceIds);
-    commandContext.getHistoricTaskInstanceManager().deleteHistoricTaskInstancesByProcessInstanceIds(processInstanceIds, false);
-    commandContext.getHistoricActivityInstanceManager().deleteHistoricActivityInstancesByProcessInstanceIds(processInstanceIds);
-    commandContext.getHistoricIncidentManager().deleteHistoricIncidentsByProcessInstanceIds(processInstanceIds);
-    commandContext.getHistoricJobLogManager().deleteHistoricJobLogsByProcessInstanceIds(processInstanceIds);
-    commandContext.getHistoricExternalTaskLogManager().deleteHistoricExternalTaskLogsByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricDetailManager().deleteHistoricDetailsByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricVariableInstanceManager().deleteHistoricVariableInstanceByProcessInstanceIds(processInstanceIds);
+      commandContext.getCommentManager().deleteCommentsByProcessInstanceIds(processInstanceIds);
+      commandContext.getAttachmentManager().deleteAttachmentsByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricTaskInstanceManager().deleteHistoricTaskInstancesByProcessInstanceIds(processInstanceIds, false);
+      commandContext.getHistoricActivityInstanceManager().deleteHistoricActivityInstancesByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricIncidentManager().deleteHistoricIncidentsByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricJobLogManager().deleteHistoricJobLogsByProcessInstanceIds(processInstanceIds);
+      commandContext.getHistoricExternalTaskLogManager().deleteHistoricExternalTaskLogsByProcessInstanceIds(processInstanceIds);
+      commandContext.getAuthorizationManager().deleteAuthorizationsByResourceIds(Resources.HISTORIC_PROCESS_INSTANCE, processInstanceIds);
 
-    commandContext.getDbEntityManager().deletePreserveOrder(HistoricProcessInstanceEntity.class, "deleteHistoricProcessInstances", processInstanceIds);
+      commandContext.getDbEntityManager().deletePreserveOrder(HistoricProcessInstanceEntity.class, "deleteHistoricProcessInstances", processInstanceIds);
+    }
   }
 
   public long findHistoricProcessInstanceCountByQueryCriteria(HistoricProcessInstanceQueryImpl historicProcessInstanceQuery) {
@@ -119,7 +121,7 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
 
   @SuppressWarnings("unchecked")
   public List<String> findHistoricProcessInstanceIdsForCleanup(Integer batchSize, int minuteFrom, int minuteTo) {
-    Map<String, Object> parameters = new HashMap<String, Object>();
+    Map<String, Object> parameters = new HashMap<>();
     parameters.put("currentTimestamp", ClockUtil.getCurrentTime());
     if (minuteTo - minuteFrom + 1 < 60) {
       parameters.put("minuteFrom", minuteFrom);
@@ -133,6 +135,12 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
   public List<String> findHistoricProcessInstanceIds(HistoricProcessInstanceQueryImpl historicProcessInstanceQuery) {
     configureQuery(historicProcessInstanceQuery);
     return (List<String>) getDbEntityManager().selectList("selectHistoricProcessInstanceIdsByQueryCriteria", historicProcessInstanceQuery);
+  }
+
+  @SuppressWarnings("unchecked")
+  public List<ImmutablePair<String, String>> findDeploymentIdMappingsByQueryCriteria(HistoricProcessInstanceQueryImpl historicProcessInstanceQuery) {
+    configureQuery(historicProcessInstanceQuery);
+    return getDbEntityManager().selectList("selectHistoricProcessInstanceDeploymentIdMappingsByQueryCriteria", historicProcessInstanceQuery);
   }
 
   @SuppressWarnings("unchecked")
@@ -191,7 +199,12 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
     commandContext.getByteArrayManager()
       .addRemovalTimeToByteArraysByRootProcessInstanceId(rootProcessInstanceId, removalTime);
 
-    Map<String, Object> parameters = new HashMap<String, Object>();
+    if (isEnableHistoricInstancePermissions()) {
+      commandContext.getAuthorizationManager()
+          .addRemovalTimeToAuthorizationsByRootProcessInstanceId(rootProcessInstanceId, removalTime);
+    }
+
+    Map<String, Object> parameters = new HashMap<>();
     parameters.put("rootProcessInstanceId", rootProcessInstanceId);
     parameters.put("removalTime", removalTime);
 
@@ -237,6 +250,11 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
 
     commandContext.getByteArrayManager()
       .addRemovalTimeToByteArraysByProcessInstanceId(processInstanceId, removalTime);
+
+    if (isEnableHistoricInstancePermissions()) {
+      commandContext.getAuthorizationManager()
+          .addRemovalTimeToAuthorizationsByProcessInstanceId(processInstanceId, removalTime);
+    }
 
     Map<String, Object> parameters = new HashMap<>();
     parameters.put("processInstanceId", processInstanceId);
@@ -311,6 +329,11 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
 
     deleteOperations.put(deleteByteArrays.getEntityType(), deleteByteArrays);
 
+    DbOperation deleteAuthorizations = commandContext.getAuthorizationManager()
+        .deleteAuthorizationsByRemovalTime(removalTime, minuteFrom, minuteTo, batchSize);
+
+    deleteOperations.put(deleteAuthorizations.getEntityType(), deleteAuthorizations);
+
     Map<String, Object> parameters = new HashMap<>();
     parameters.put("removalTime", removalTime);
     if (minuteTo - minuteFrom + 1 < 60) {
@@ -326,6 +349,11 @@ public class HistoricProcessInstanceManager extends AbstractHistoricManager {
     deleteOperations.put(deleteProcessInstances.getEntityType(), deleteProcessInstances);
 
     return deleteOperations;
+  }
+
+  protected boolean isEnableHistoricInstancePermissions() {
+    return Context.getProcessEngineConfiguration()
+        .isEnableHistoricInstancePermissions();
   }
 
 }
